rank to int, rename name -> user

+++ domain/nba/nba_wins_pool_service.py
@@ -149,7 +149,7 @@ class NbaWinsPoolService:
             .sort_values(by=["wins", "losses"], ascending=[False, True], ignore_index=True)
         )
 
-        leaderboard_df["rank"] = leaderboard_df["wins"].rank(method="min", ascending=False)
+        leaderboard_df["rank"] = leaderboard_df["wins"].rank(method="min", ascending=False).astype(int)
 
         return leaderboard_df
